Refactoring moved internal/testdata to test/testdata.
