+++ test/integration_tests/metadata_resource_test.go
@@ -13,8 +13,8 @@ import (
 	"context"
 	"testing"
 
-	"github.com/conductor-sdk/conductor-go/internal/testdata"
 	"github.com/conductor-sdk/conductor-go/sdk/model"
+	"github.com/conductor-sdk/conductor-go/test/testdata"
 	"github.com/stretchr/testify/assert"
 )
