+++ test/integration_tests/task_resource_test.go
@@ -14,10 +14,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/conductor-sdk/conductor-go/internal/testdata"
 	"github.com/conductor-sdk/conductor-go/sdk/model"
 	"github.com/conductor-sdk/conductor-go/sdk/workflow"
 	"github.com/conductor-sdk/conductor-go/test/common"
+	"github.com/conductor-sdk/conductor-go/test/testdata"
 )
 
 func TestUpdateTaskRefByName(t *testing.T) {
