+++ test/integration_tests/workflow_def_test.go
@@ -14,10 +14,10 @@ import (
 	"os"
 	"testing"
 
-	"github.com/conductor-sdk/conductor-go/internal/testdata"
 	"github.com/conductor-sdk/conductor-go/sdk/model"
 	"github.com/conductor-sdk/conductor-go/sdk/workflow"
 	"github.com/conductor-sdk/conductor-go/test/common"
+	"github.com/conductor-sdk/conductor-go/test/testdata"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/assert"
 )
