+++ test/testdata/kitchensink.go
