+++ test/unit_tests/workflow_def_test.go
@@ -3,10 +3,12 @@ package unit_tests
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/conductor-sdk/conductor-go/internal/testdata"
-	"github.com/conductor-sdk/conductor-go/sdk/workflow"
-	"github.com/stretchr/testify/assert"
+
 	"testing"
+
+	"github.com/conductor-sdk/conductor-go/sdk/workflow"
+	"github.com/conductor-sdk/conductor-go/test/testdata"
+	"github.com/stretchr/testify/assert"
 )
 
 func TestRetrySettings(t *testing.T) {
